+++ NEWS.md
@@ -3,6 +3,8 @@
 * Format time using commonyears + days + hh:mm:ss.sss s.
 * Modify @(unit, text) operator to return text instead of cstring.
 * Import definitions.units from GNU units 2.17 with 13 new units.
+* Remove bison 2 compat mode, but keep pregenerated files for jessie and
+  trusty.
 
 6.0: Mar 7, 2018
 ----------------