+++ NEWS.md
@@ -7,6 +7,7 @@
 * Remove bison 2 compat mode, but keep pregenerated files for jessie and
   trusty.
 * Support send/receive via the binary protocol.
+* Support Infinity and NaN.
 
 6.0: Mar 7, 2018
 ----------------