+++ NEWS.md
@@ -1,6 +1,7 @@
 7.0: Jul ?, 2018
 ----------------
 * Format time using commonyears + days + hh:mm:ss.sss s.
+* Avoid using Gs for Gigaseconds, that's actually gauss.
 * Modify @(unit, text) operator to return text instead of cstring.
 * Import definitions.units from GNU units 2.17 with 13 new units.
 * Add π to known units.